% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions (\code{gibbsR} and \code{vaccR}) and Cpp functions (\code{gibbsC} and \code{vaccC}).
}
\examples{
\dontrun{
microbenchmark(sumtwo(1,2),timesTwo(100))
}
}
