% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{gaussian_probability}
\alias{gaussian_probability}
\title{Title}
\usage{
gaussian_probability(x, mu, sigma)
}
\arguments{
\item{x}{a instance}

\item{mu}{mean}

\item{sigma}{std}
}
\value{
probabilty denstiny
}
\description{
Title
}
