% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris_trim.R
\docType{data}
\name{iris_trim}
\alias{iris_trim}
\title{Part Of Edgar Anderson's Iris Data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{\code{Sepal.Length}}{double The unit is cm}
\item{\code{Sepal.Width}}{double The unit is cm}
\item{\code{Species}}{double 1 and -1 represent two types}
}
}
\source{
Fisher, R. A. (1936) The use of multiple measurements in taxonomic problems. Annals of Eugenics, 7, Part II, 179–188.
}
\usage{
iris_trim
}
\description{
Intercept the first two features of the first 100 rows of the iris data set, and mark their types with two categories
}
\keyword{datasets}
