% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perceptronC}
\alias{perceptronC}
\title{perceptronC}
\usage{
perceptronC(w, b, l_rate, feature, target)
}
\arguments{
\item{w}{wight vector}

\item{b}{bias}

\item{l_rate}{learning rate}

\item{feature}{characteristic matrix}

\item{target}{category}
}
\value{
parameters of hyperplane
}
\description{
divide linearly separable datasets into two categories
}
\examples{
\dontrun{
data(iris_trim)
w = c(0,0)
b=0
f = iris_trim[,c(1,2)]
t = iris_trim[,5]
l_rate =0.001
perceptronC(w,b,l_rate,f,t)
}
}
