% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{perceptronR}
\alias{perceptronR}
\title{Title}
\usage{
perceptronR(w, b, l_rate, feature, target)
}
\arguments{
\item{w}{wight vector}

\item{b}{bias}

\item{l_rate}{learning rate}

\item{feature}{characteristic matrix}

\item{target}{category}
}
\value{
parameters of hyperplane
}
\description{
Title
}
\examples{
\dontrun{
library(StatComp20042)
w = c(0,0)
b=0
f = iris_trim[,c(1,2)]
t = iris_trim[,3]
l_rate =0.001
perceptronR(w,b,l_rate,f,t)
}
}
