\name{StatComp20042-package}
\alias{StatComp20042-package}
\docType{package}
\title{
  This package provides several functions for reproducing several classical algorithms in machine learning.
}
\description{
  R,Rcpp version of perceptron algorithm and R version of gaussian naive bayes algorithm are provided in this package (\code{perceptronC},\code{perceptronR},\code{gnb_fit},\code{gnb_pre})
  R and Rcpp versions of gibbs sampler and vaac function are provided in this package (\code{gibbsR} and \code{vaccR} versus \code{gibbsC} and \code{vaccC})
}
\author{
Weijie Cui

Maintainer: Weijie Cui <can@mail.ustc.edu.cn>
}
\keyword{ package }
