% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timesTwo}
\alias{timesTwo}
\title{timesTwo}
\usage{
timesTwo(x)
}
\arguments{
\item{x}{the number of between-sample random numbers}
}
\value{
a random sample of size a
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
timesTwo(x)
}
}
