% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{train_test_split}
\alias{train_test_split}
\title{Split data to training data set and test data set}
\usage{
train_test_split(raw_data, train_size, raw_seed)
}
\arguments{
\item{raw_data}{data to be splited}

\item{train_size}{proportion of training set}

\item{raw_seed}{random seed}
}
\value{
training set data and test set data
}
\description{
Split data to training data set and test data set
}
\examples{
\dontrun{
train_test_split(raw_data=iris_trim, train_size=0.8, raw_seed=112)
}
}
