% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{gnb_fit}
\alias{gnb_fit}
\title{Fit a gaussian naive bayes model.}
\usage{
gnb_fit(train)
}
\arguments{
\item{train}{A matrix composed of column vectors. The last column is the category and the former columns are features.}
}
\value{
A list composed of prior distribution,a matrix of mean values and a matrix of std values.
}
\description{
Fit a gaussian naive bayes model.
}
\examples{
\dontrun{
data(iris_trim)
train_and_test <- train_test_split(raw_data=iris_trim, train_size=0.8, raw_seed=112)
train <- train_and_test[[1]]
test <- train_and_test[[2]]
gnb_list = gnb_fit(train)
print(gnb_list)
 }
}
