% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{gnb_pre}
\alias{gnb_pre}
\title{Predict category of test data using gaussian naive bayes model.}
\usage{
gnb_pre(test_array, gnb_list)
}
\arguments{
\item{test_array}{A feature vector to be predicted}

\item{gnb_list}{the return value of gnb_fit model}
}
\value{
the category of test_array
}
\description{
Predict category of test data using gaussian naive bayes model.
}
