% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris_trim.R
\docType{data}
\name{iris_trim}
\alias{iris_trim}
\title{Part Of Edgar Anderson's Iris Data}
\format{
A data frame with 100 rows and 5 variables:
\describe{
  \item{\code{Sepal.Length}}{The unit is cm}
  \item{\code{Sepal.Width}}{The unit is cm}
  \item{\code{Petal.Length}}{The unit is cm}
  \item{\code{Petal.Width}}{The unit is cm}
  \item{\code{category}}{1 and -1 represent two types}
}
}
\source{
Fisher, R. A. (1936) The use of multiple measurements in taxonomic problems. Annals of Eugenics, 7, Part II, 179–188.
}
\usage{
iris_trim
}
\description{
Intercept the first 100 rows of the iris data set, and mark their types with two categories
}
\keyword{datasets}
