% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{perceptronR}
\alias{perceptronR}
\title{Computing hyperplanes for binary classification problems}
\usage{
perceptronR(w, b, l_rate, feature, c)
}
\arguments{
\item{w}{weight vector}

\item{b}{bias}

\item{l_rate}{learning rate}

\item{feature}{character matrix}

\item{c}{category}
}
\value{
parameters of hyperplane
}
\description{
Computing hyperplanes for binary classification problems
}
\examples{
\dontrun{
data(iris_trim)
w = c(0,0)
b=0
f = iris_trim[,c(1,2)]
t = iris_trim[,5]
l_rate =0.001
perceptronR(w,b,l_rate,f,t)
}
}
