\name{StatComp20042-package}
\alias{StatComp20042-package}
\docType{package}
\title{
  This package reproduces several machine learning classical algorithms.
}
\description{
  R,Rcpp version of perceptron algorithm and R version of gaussian naive bayes algorithm are provided in this package (\code{perceptronC},\code{perceptronR},\code{gnb_fit},\code{gnb_pre})
}
\author{
Weijie Cui

Maintainer: Weijie Cui <can@mail.ustc.edu.cn>
}
\keyword{ package }
